+++ erdjs/src/smartcontracts/codec/boolean.ts
@@ -1,0 +1,59 @@
+import * as errors from "../../errors";
+import { BooleanValue } from "../typesystem";
+
+export class BooleanBinaryCodec {
+    private static readonly TRUE: number = 0x01;
+    private static readonly FALSE: number = 0x00;
+
+    /**
+    * Reads and decodes a BooleanValue from a given buffer, 
+    * with respect to: {@link https://docs.elrond.com/developers/developer-reference/the-elrond-serialization-format | The Elrond Serialization Format}. 
+    * 
+    * @param buffer the input buffer
+    */
+    decodeNested(buffer: Buffer): [BooleanValue, number] {
+        // We don't not check the size of the buffer, we just read the first byte.
+
+        let byte = buffer.readUInt8();
+        return [new BooleanValue(byte == BooleanBinaryCodec.TRUE), 1];
+    }
+
+    /**
+     * Reads and decodes a BooleanValue from a given buffer, 
+     * with respect to: {@link https://docs.elrond.com/developers/developer-reference/the-elrond-serialization-format | The Elrond Serialization Format}. 
+     * 
+     * @param buffer the input buffer
+     */
+    decodeTopLevel(buffer: Buffer): BooleanValue {
+        if (buffer.length > 1) {
+            throw new errors.ErrInvalidArgument("buffer", buffer, "should be a buffer of size <= 1");
+        }
+
+        let firstByte = buffer[0];
+        return new BooleanValue(firstByte == BooleanBinaryCodec.TRUE);
+    }
+
+    /**
+     * Encodes a BooleanValue to a buffer, 
+     * with respect to: {@link https://docs.elrond.com/developers/developer-reference/the-elrond-serialization-format | The Elrond Serialization Format}. 
+     */
+    encodeNested(primitive: BooleanValue): Buffer {
+        if (primitive.isTrue()) {
+            return Buffer.from([BooleanBinaryCodec.TRUE]);
+        }
+
+        return Buffer.from([BooleanBinaryCodec.FALSE]);
+    }
+
+    /**
+     * Encodes a BooleanValue to a buffer, 
+     * with respect to: {@link https://docs.elrond.com/developers/developer-reference/the-elrond-serialization-format | The Elrond Serialization Format}. 
+     */
+    encodeTopLevel(primitive: BooleanValue): Buffer {
+        if (primitive.isTrue()) {
+            return Buffer.from([BooleanBinaryCodec.TRUE]);
+        }
+
+        return Buffer.from([]);
+    }
+}
