+++ erdjs/src/smartcontracts/codec/utils.ts
@@ -1,0 +1,105 @@
+
+/**
+ * Returns whether the most significant bit of a given byte (within a buffer) is 1.
+ * @param buffer the buffer to test
+ * @param byteIndex the index of the byte to test
+ */
+export function isMbsOne(buffer: Buffer, byteIndex: number = 0): boolean {
+    let byte = buffer[byteIndex];
+    let bit = byte >> 7;
+    let isSet = bit % 2 == 1;
+    return isSet;
+}
+
+/**
+ * Returns whether the most significant bit of a given byte (within a buffer) is 0.
+ * @param buffer the buffer to test
+ * @param byteIndex the index of the byte to test
+ */
+export function isMbsZero(buffer: Buffer, byteIndex: number = 0): boolean {
+    return !isMbsOne(buffer, byteIndex);
+}
+
+export function cloneBuffer(buffer: Buffer) {
+    let clone = Buffer.alloc(buffer.length);
+    buffer.copy(clone);
+    return clone;
+}
+
+export function bufferToBigInt(buffer: Buffer): bigint {
+    // Currently, in JavaScript, this is the feasible way to achieve reliable, arbitrary-size Buffer to BigInt conversion.
+    let hex = buffer.toString("hex");
+    let value = BigInt(`0x${hex}`);
+    return value;
+}
+
+export function bigIntToBuffer(value: bigint): Buffer {
+    // Currently, in JavaScript, this is the feasible way to achieve reliable, arbitrary-size BigInt to Buffer conversion.
+    let hex = getHexMagnitudeOfBigInt(value);
+    let buffer = Buffer.from(hex, "hex");
+    return buffer;
+}
+
+export function getHexMagnitudeOfBigInt(value: bigint): string {
+    if (!value) {
+        return "";
+    }
+
+    if (value < BigInt(0)) {
+        value = value * BigInt(-1);
+    }
+
+    let hex = value.toString(16);
+    let padding = "0";
+
+    if (hex.length % 2 == 1) {
+        hex = padding + hex;
+    }
+
+    return hex;
+}
+
+export function flipBufferBitsInPlace(buffer: Buffer) {
+    for (let i = 0; i < buffer.length; i++) {
+        buffer[i] = ~buffer[i];
+    }
+}
+
+export function prependByteToBuffer(buffer: Buffer, byte: number) {
+    return Buffer.concat([Buffer.from([byte]), buffer]);
+}
+
+
+/**
+ * Discards the leading bytes that are merely a padding of the leading sign bit (but keeps the payload).
+ * @param buffer A number, represented as a sequence of bytes (big-endian)
+ */
+export function discardSuperfluousBytesInTwosComplement(buffer: Buffer): Buffer {
+    let isNegative = isMbsOne(buffer, 0);
+    let signPadding: number = isNegative ? 0xFF : 0x00;
+
+    let index;
+    for (index = 0; index < buffer.length - 1; index++) {
+        let isPaddingByte = buffer[index] == signPadding;
+        let hasSignBitOnNextByte = isMbsOne(buffer, index + 1) === isNegative;
+        if (isPaddingByte && hasSignBitOnNextByte) {
+            continue;
+        }
+
+        break;
+    }
+
+    return buffer.slice(index);
+}
+
+/**
+ * Discards the leading zero bytes.
+ * @param buffer A number, represented as a sequence of bytes (big-endian)
+ */
+export function discardSuperfluousZeroBytes(buffer: Buffer): Buffer {
+    let index;
+    for (index = 0; index < buffer.length && buffer[index] == 0; index++) {
+    }
+
+    return buffer.slice(index);
+}
